db_options.rs: deprecate set_ignore_range_deletions (RocksDB 10.2.1) (#1000)
RocksDB 10.2.1 is marking this API as deprecated. The upstream
documentation has changed to state the following. Update rustdoc to
match, and mark with #[deprecated]. Currently rust-rocksdb is not
using this version of RocksDB, but presumably it will eventually, so
we might as well mark this as deprecated now.

Upstream RocksDB comment:

// DEPRECATED: This option might be removed in a future release.
// There should be no noticeable performance difference whether this option
// is turned on or off when a DB does not use DeleteRange().

https://github.com/facebook/rocksdb/blob/v10.2.1/include/rocksdb/options.h#L1837